Add Qodana GitHub Actions workflow
Added a new file `qodana.yml` to the `.github/workflows` directory that contains a new GitHub Actions workflow for Qodana. This was done to automate code quality checks and apply fixes for every push event to the main branch as well as on pull requests. The Qodana scan runs on Ubuntu latest and enables auto-fixing of detected issues.